Extract isHierarchicalPostType to new file
