+++ lib/wordpress/_config/postTypes.js
@@ -16,14 +16,3 @@ export const postTypes = {
 
 // Define hierarchical post types.
 export const hierarchicalPostTypes = ['page']
-
-/**
- * Check if post type is hierarchical.
- *
- * @author WebDevStudios
- * @param {string} postType WP post type.
- * @return {boolean} Whether provided post type is hierarchical.
- */
-export function isHierarchicalPostType(postType) {
-  return hierarchicalPostTypes.includes(postType)
-}
