+++ lib/wordpress/_global/getPostTypeStaticPaths.js
@@ -1,9 +1,7 @@
+import isHierarchicalPostType from '@/functions/wordpress/posts/isHierarchicalPostType'
 import isValidPostType from '@/functions/wordpress/posts/isValidPostType'
 import {initializeWpApollo} from '@/lib/wordpress/connector'
-import {
-  isHierarchicalPostType,
-  postTypes
-} from '@/lib/wordpress/_config/postTypes'
+import {postTypes} from '@/lib/wordpress/_config/postTypes'
 import {gql} from '@apollo/client'
 
 /**
